refactor proposal narrative review router
